# Conflicts:
#	src/main/java/com/internship/tmontica/config/WebConfig.java

+++ src/main/java/com/internship/tmontica/user/UserService.java
@@ -1,11 +1,15 @@
 package com.internship.tmontica.user;
 
 import com.internship.tmontica.security.JwtServiceImpl;
+import com.internship.tmontica.security.exception.UnauthorizedException;
 import com.internship.tmontica.user.exception.*;
 import com.internship.tmontica.user.model.request.*;
+import com.internship.tmontica.user.model.response.UserFindIdRespDTO;
 import com.internship.tmontica.user.model.response.UserInfoRespDTO;
 import com.internship.tmontica.security.AuthenticationKey;
 import com.internship.tmontica.user.model.response.UserSignInRespDTO;
+import com.internship.tmontica.user.model.response.UserTokenInfoDTO;
+import com.internship.tmontica.util.JsonUtil;
 import com.internship.tmontica.util.UserConfigValueName;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
@@ -27,7 +31,7 @@ public class UserService { //implements UserDetail
     private final ModelMapper modelMapper;
     private final MailSender sender;
     private final JwtServiceImpl jwtService;
-    private static final String ID_AUTH_CODE = "authCode";
+
 
     private enum MailOption {
         findId, findPassword
@@ -37,6 +41,7 @@ public class UserService { //implements UserDetail
     public boolean signUp(UserSignUpReqDTO userSignUpReqDTO){
         //userSignUpReqDTO.setPassword(passwordEncoder.encode(user.getPassword())); need passwordEnconding
         checkUserIdDuplicatedException(userSignUpReqDTO.getId());
+        checkPasswordMismatchException(userSignUpReqDTO.getPassword(), userSignUpReqDTO.getPasswordCheck());
         setRole(userSignUpReqDTO);
         User user = modelMapper.map(userSignUpReqDTO, User.class);
         return userDao.addUser(user) > 0;
@@ -46,11 +51,12 @@ public class UserService { //implements UserDetail
         String role = userSignUpReqDTO.getRole();
         if(role == null || role.equals(UserRole.USER.toString())){
             userSignUpReqDTO.setRole("USER");
+            return;
         } else if (role.equals(UserRole.ADMIN.toString())){
-            userSignUpReqDTO.setRole("ADMIN");
-        } else {
-            throw new InvalidUserRoleException();
+            return;
         }
+
+        throw new InvalidUserRoleException();
     }
 
     @Transactional(readOnly = true)
@@ -72,28 +78,22 @@ public class UserService { //implements UserDetail
                 userDao.getUserByUserId(userSignInReqDTO.getId()).getPassword());
     }
 
-    // USER 값 선택하게 수정
+    @Transactional(readOnly = true)
      public UserSignInRespDTO makeJwtToken(UserSignInReqDTO userSignInReqDTO){
 
-        return new UserSignInRespDTO("true", jwtService.getToken(makeTokenUserWithRole(userSignInReqDTO.getId(),
-                "USER")));
-//        (UserConfigValueName.JWT_TOKEN_HEADER_KEY,
-//                jwtService.getToken(makeTokenUserWithRole(userSignInReqDTO.getId(),
-//                        "USER")));
+        return new UserSignInRespDTO(jwtService.getToken(makeTokenUser(userSignInReqDTO.getId())));
     }
 
-    private User makeTokenUserWithRole(String id, String role){
+    private UserTokenInfoDTO makeTokenUser(String id){
 
-        User user = new User();
-        user.setId(id);
-        user.setRole(role);
-        return user;
+        User user = userDao.getUserByUserId(id);
+        return modelMapper.map(user, UserTokenInfoDTO.class);
     }
 
     @Transactional(readOnly = true)
     public void checkPassword(UserCheckPasswordReqDTO userCheckPasswordReqDTO){
 
-        String userId = jwtService.getUserInfo("userId");
+        String userId = JsonUtil.getJsonElementValue(jwtService.getUserInfo("userInfo"), "id");
         checkMissingSessionUserIdException(userId);
         checkPasswordMismatchException(userCheckPasswordReqDTO.getPassword(),
                 userDao.getUserByUserId(userId).getPassword());
@@ -102,7 +102,7 @@ public class UserService { //implements UserDetail
     @Transactional
     public boolean changePassword(UserChangePasswordReqDTO userChangePasswordReqDTO){
 
-        String userId = jwtService.getUserInfo("userId");
+        String userId = JsonUtil.getJsonElementValue(jwtService.getUserInfo("userInfo"), "id");
         checkMissingSessionUserIdException(userId);
         checkPasswordMismatchException(userChangePasswordReqDTO.getNewPassword(), userChangePasswordReqDTO.getNewPasswordCheck());
         User user = modelMapper.map(userChangePasswordReqDTO, User.class);
@@ -120,6 +120,10 @@ public class UserService { //implements UserDetail
     @Transactional
     public boolean withDrawUser(String id){
 
+        String userId = JsonUtil.getJsonElementValue(jwtService.getUserInfo("userInfo"), "id");
+        if(!userId.equals(id)){
+            throw new UnauthorizedException();
+        }
         checkUserIdNotFoundException(id);
         return userDao.deleteUser(id) > 0;
     }
@@ -127,10 +131,13 @@ public class UserService { //implements UserDetail
     @Transactional(readOnly = true)
     public boolean sendUserId(String email, HttpSession httpSession) throws MailSendException{
         User user = userDao.getUserByEmail(email);
-        checkUserIdNotFoundException(user.getId());
+        if(user == null){
+            return false;
+        }
         AuthenticationKey authenticationKey = new AuthenticationKey();
         sender.send(sendMail(user, MailOption.findId, authenticationKey.getAuthenticationKey()));
-        httpSession.setAttribute(ID_AUTH_CODE, authenticationKey.getAuthenticationKey());
+        httpSession.setAttribute(UserConfigValueName.ID_AUTH_CODE, authenticationKey.getAuthenticationKey());
+        httpSession.setAttribute(UserConfigValueName.FIND_ID_SESSION_ATTRIBUTE_NAME, user.getId());
         return true;
     }
 
@@ -143,10 +150,15 @@ public class UserService { //implements UserDetail
     }
 
     @Transactional
-    public boolean checkAuthCode(UserFindIdReqDTO userFindIdReqDTO, HttpSession session){
+    public UserFindIdRespDTO checkAuthCode(UserFindIdReqDTO userFindIdReqDTO, HttpSession session){
 
-        String sessionAttribute = checkSessionAttribute(session, ID_AUTH_CODE);
-        return userFindIdReqDTO.getAuthCode().equals(sessionAttribute);
+        String authCode = checkSessionAttribute(session, UserConfigValueName.ID_AUTH_CODE);
+        if(userFindIdReqDTO.getAuthCode().equals(authCode)){
+            String findId = checkSessionAttribute(session, UserConfigValueName.FIND_ID_SESSION_ATTRIBUTE_NAME);
+            return new UserFindIdRespDTO(findId, true);
+        }
+
+        return new UserFindIdRespDTO("",false );
     }
 
     private String checkSessionAttribute(HttpSession session, String key){
@@ -236,5 +248,4 @@ public class UserService { //implements UserDetail
 //        return new org.springframework.security.core.userdetails.User(user.getId(), user.getPassword(),authorities);
 //
 //    }
-
 }
